refactor(bridge): extract forwarding cache
