+++ whatsrust/lib/main_test.go
@@ -102,20 +102,6 @@ func TestViewOnceWrappersBecomeUnavailablePlaceholders(t *testing.T) {
 	}
 }
 
-func TestViewOnceMessagesAreNotCachedForForwarding(t *testing.T) {
-	resetForwardedSourcesForTest()
-	t.Cleanup(resetForwardedSourcesForTest)
-	chat := types.NewJID("chat", types.DefaultUserServer)
-	info := types.MessageInfo{MessageSource: types.MessageSource{Chat: chat, Sender: chat}, ID: "view-once"}
-	message := &waE2E.Message{ViewOnceMessage: &waE2E.FutureProofMessage{Message: &waE2E.Message{ImageMessage: &waE2E.ImageMessage{}}}}
-
-	cacheForwardSource(info, message)
-
-	if len(forwardedSources.entries) != 0 {
-		t.Fatal("view-once media must not be cached for forwarding")
-	}
-}
-
 func TestOrdinaryAndEphemeralMessagesRemainAvailable(t *testing.T) {
 	ordinary := &waE2E.Message{Conversation: stringPointer("ordinary")}
 	if got, unavailable := unavailableViewOnceMessage(ordinary); unavailable || got != ordinary {
@@ -1428,34 +1414,6 @@ func TestNewForwardRequestRejectsBroadcastAndInvalidDestinations(t *testing.T) {
 	request, err := newForwardRequest(chat.String(), sender.String(), "message", []string{chat.String()})
 	if err != nil || len(request.destinations) != 1 {
 		t.Fatalf("forward to source chat must be allowed: request=%#v err=%v", request, err)
-	}
-}
-
-func TestForwardSourceCacheEvictsAndInvalidatesDeletedSources(t *testing.T) {
-	resetForwardedSourcesForTest()
-	t.Cleanup(resetForwardedSourcesForTest)
-	chat := types.NewJID("source", types.DefaultUserServer)
-	sender := types.NewJID("sender", types.DefaultUserServer)
-	for index := 0; index <= maxForwardSources; index++ {
-		id := types.MessageID(fmt.Sprintf("message-%d", index))
-		cacheForwardSource(types.MessageInfo{MessageSource: types.MessageSource{Chat: chat, Sender: sender}, ID: id}, &waE2E.Message{Conversation: stringPointer("payload")})
-	}
-	first := forwardSourceKey(chat, sender, "message-0")
-	last := forwardSourceKey(chat, sender, types.MessageID(fmt.Sprintf("message-%d", maxForwardSources)))
-	forwardedSources.mu.Lock()
-	_, firstExists := forwardedSources.entries[first]
-	_, lastExists := forwardedSources.entries[last]
-	entryCount := len(forwardedSources.entries)
-	forwardedSources.mu.Unlock()
-	if firstExists || !lastExists || entryCount != maxForwardSources {
-		t.Fatalf("cache eviction failed: first=%t last=%t count=%d", firstExists, lastExists, entryCount)
-	}
-	removeForwardSources(chat.String(), "message-"+fmt.Sprint(maxForwardSources))
-	forwardedSources.mu.Lock()
-	_, lastExists = forwardedSources.entries[last]
-	forwardedSources.mu.Unlock()
-	if lastExists {
-		t.Fatal("deleted source remained forwardable in cache")
 	}
 }
